+++ NEWS.md
@@ -1,2 +1,6 @@
+# seeker 0.0.0.9067
+* Changed `species` to `organism` for consistency with NCBI and other packages.
+* Added checking for parallel backend.
+
 # seeker 0.0.0.9062
 * Fixed syntax to adhere to lab code style.